+++ fsl_mrs/tests/test_scripts_fsl_mrsi.py
@@ -74,6 +74,7 @@ def test_fsl_mrsi(tmp_path):
 
     assert (tmp_path / 'fit_out/concs/raw/NAA.nii.gz').exists()
     assert (tmp_path / 'fit_out/concs/molality/NAA.nii.gz').exists()
+    assert not (tmp_path / 'fit_out/concs/old_WRONG_molality/NAA.nii.gz').exists()
     assert (tmp_path / 'fit_out/uncertainties/NAA_sd.nii.gz').exists()
     assert (tmp_path / 'fit_out/qc/NAA_snr.nii.gz').exists()
     assert (tmp_path / 'fit_out/qc/fit_failed.nii.gz').exists()
@@ -248,6 +249,7 @@ def test_alt_ref(tmp_path):
                            data['seg_gm'],
                            data['seg_csf'],
                            '--output_correlations',
+                           '--export_old_wrong_molality',
                            '--overwrite',
                            '--combine', 'Cr', 'PCr',
                            '--wref_metabolite', 'PCho', 'GPC',
@@ -262,6 +264,7 @@ def test_alt_ref(tmp_path):
 
     assert (tmp_path / 'fit_out/concs/raw/NAA.nii.gz').exists()
     assert (tmp_path / 'fit_out/concs/molality/NAA.nii.gz').exists()
+    assert (tmp_path / 'fit_out/concs/old_WRONG_molality/NAA.nii.gz').exists()
     assert (tmp_path / 'fit_out/uncertainties/NAA_sd.nii.gz').exists()
     assert (tmp_path / 'fit_out/qc/NAA_snr.nii.gz').exists()
     assert (tmp_path / 'fit_out/fit/fit.nii.gz').exists()